Fix Regex
Know that whenever you made a regex that looks like it works: 
-> IT DOESN'T

You can be sure that there is something you missed.

The s in the beginning of speed and signal where picked up as a [s] (strikethrough) tag from bbcode (which get's removed before checking for the dialogic tags. So it didn't consider these as tags.