Add check for correct text length
This is a simple fix that prevents the [img] bbcode from breaking the game (#714). It's not perfect, because it means after the image all checks for a specific character will be off. That's because the image counts as a "visible_character" but has no  character in the $RichTextLabel.text property.