Adding history functionality and interface (#625)
Adds a new section in the settings where you can enable a history view. 
The history is a window that shown and hidden (either with build in buttons or via a function). It consists of a panel and a VBox. For each event that is logged (currently text, question, character) there will be a HistoryRow instanced that displays the content of the event. If the event had audio, this can be replayed. Many more things can be customized and for more the user should customize the HistoryRow scene. 

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>
Co-authored-by: Emilio Coppola <coppolaemilio@protonmail.com>
Co-authored-by: Jowan-Spooner <42868150+Jowan-Spooner@users.noreply.github.com>