Add new dialogic_default_action and fix bug #736 (#795)
There is now a dialogic_default_action automatically created in the InputMap. It is now also the new default. 
Unfortunately you have to force the editor to update by reloading the project or adding another action to able to see this action. But this is only a problem the first time you open the project (or if you delete the action).

Also the editors internal InputMap will no longer be messed up when setting the action key (#736).

Co-authored-by: Emilio Coppola <coppolaemilio@protonmail.com>