Remove warning logic check for "["
I was mistaken, this is not how it works. Instead you type the name of the value definition WITHOUT anything around it. Because of this, there is now way to know if you want to use a string or a definition. Thus I removed the warning completely from the Condition picker.

On the SetValue event however, using values isn't implemented at all, so I removed the check for "[" there as well.