Small UI updates (#804)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Cleanup and Simplification of the MasterTree.gd

* Auto stash before merge of "main" and "upstream/main"

* Remove more Engine.get_singleton()'s

:(

* Make the DocsTreeHelper ignore the mouse

It was blocking a small part of the MasterTree. This should fix #414

* VoiceSupport Feature Fix: Make sure no default bus is needed

There was a little annyoing warning popping up a lot.

* Small Documentation fix

* Add event documentation

Adds a basic documentation of
- Leave Event
- Question Event
- Choice Event
- End Event
- Condition Event
- Set Value Event

* Add the new documentations to the events

You now have these little buttons on the right of the events that instantly lead to the documentation.

* Check for complete dialogic_state information in start_from_save

Apparently you can manage to save dialog_data that is empty. Maybe it has to do with the Autosave on timeline end? Need to investigate this. But this thing adds a check if the data is valid.

* Changing the start_from_save function once again

You now need to have "/" if you want to use the currently loaded state (mostly after importing). 
Lot's of more checks added.

* Replace a get_tree().create_timer() with the DialogicTimer

Only instance of this making sense I found. Fix for #556

* Add/Fix documentation links of TextEvent and CharacterEvent

* Revert "Add/Fix documentation links of TextEvent and CharacterEvent"

This reverts commit e1e7ecb21f71437389fd327e47487a5d75bcdd62.

* Small UI updates

Character Editor:
- enabled wrap on the description box
- made character name and display name fiels expand if the text get's to long
- updated the theme picker style to match the resource pickers in the rest of the UI
- added the search icon to the search bar and enabled the clear button

Events:
- Change the default text of the glossary picker to "Select Glossary Item" instead ot "Resource"
- Made sure that if a glossary is deleted it won't result in an error, but the event will reset
- Changed the "warning logic" in the condition picker to account for the empty default value
- Change the warning logic of condition and set value to not show the warning if a value is entered.

* Remove warning logic check for "["

I was mistaken, this is not how it works. Instead you type the name of the value definition WITHOUT anything around it. Because of this, there is now way to know if you want to use a string or a definition. Thus I removed the warning completely from the Condition picker.

On the SetValue event however, using values isn't implemented at all, so I removed the check for "[" there as well.