Implement focus() function for EventParts and use it on Text Event
This new function (focus()) is called when an event is newly created on all the direct EventParts of the event (body/header). This allows them to do anything they would like. It' s not called when the timeline is loaded. 

This PR also implements this on the TextAndVoicePicker and it's sub-event-part the TextEditor to make it focus the TextEdit when you create the event. This fixes #811