Make sure fade in doesn't mess with dialogs loaded from save (#816)
Discovered this bug while working on the VN-template. When you loaded dialog it will first start the "[No Timeline Specified]" stuff, because the load_from_save() thing waits until "_ready". Then it will overwrite this. But because the fade was triggered before it would switch back and show the "[No Timeline Specified]" again.