Add character Resource and change extensions
Adds a character resource as well as a CharcacterResourceLoader and a CharacterResourceSaver. I'm not yet sure in what format it would be best to save the character, maybe also in a way that is easy to edit from the outside. These resources are stored as '.dch' files.

I've changed the extension of the timeline resource to '.dtl'. Thus I had to replace the test timelines.