Implement Character into editor and Text event
You can now create new character resources with a button from the toolbar. There isn't really an character editor yet, although there is a placeholder node managing the creation.

To simplify the code I moved the EditorFileDialog to be a child of the EditorView so that it can be used by all the editors, not just the timeline editor.

The text event now uses the character resource. Thus the import and export had to be ajusted a bit. When importing, the text event will GUESS the resource, currently based on it's file name. We could change this to account for the actual name property. The guessing code is in the DialogicUtil, so that other events can use it easily as well. 

There has also been a developement on the ResourcePicker (as the character was a string previously). The resource picker allows to search and shows suggestions (filtered only by the file name currently, we might want to account for display name as well). Hitting enter or clicking somewhere else will act as if the first suggestion was selected. You can also drag and drop resources on the picker.