Add more events to importer/loader
- Moved all currently usable events into addons/dialogic/Resources/Events (and adjusted all references to use this path)

- Added a Comment event (starts with a # in the text format)
- Added an initial version of a Character event to expand the event builder function:
     -> you can now create selectors in the event
     -  there is also a probably working number selector for the events
     - In the text format the Character event currently starts with either "Join", "Leave" or "Update" followed by a character name. Other settings aren't yet implemented.

- Created a "default_event" in addons/dialogic/Resources. This is like a blueprint just containing all necessary methods and some explenation of what they do. This should make it easier to create new events, but we'll see.

- All events now have another static method called `is_valid_event_string` that should return true if the given string should be interpreted as event the method was called on. 
     - the text event always returns true but is asked last
     - the comment event for example checks if the string starts with a # 

Maybe we'll want to change the "one line per event" approach of saving, idk, but until we do, I've decided to convert all \n breaks into <br> characters on save and vice versa on load, so we can have multiple lines in a text event. This isn't nice to write in the text format tho, so who knows how it will end up.

Also renamed all event scripts to start with 'event_', cause then they are in one place in the scripts list and it's easier to see, what they are (in contrast to just text.gd).