Improvements (including Characters, Resource Picker, Changed Extensions) (#901)
* Some cleanup + saved marking

The current resource now has a (*) behind it's name in the toolbar if it has been changed.

* Add character Resource and change extensions

Adds a character resource as well as a CharcacterResourceLoader and a CharacterResourceSaver. I'm not yet sure in what format it would be best to save the character, maybe also in a way that is easy to edit from the outside. These resources are stored as '.dch' files.

I've changed the extension of the timeline resource to '.dtl'. Thus I had to replace the test timelines.

* Implement Character into editor and Text event

You can now create new character resources with a button from the toolbar. There isn't really an character editor yet, although there is a placeholder node managing the creation.

To simplify the code I moved the EditorFileDialog to be a child of the EditorView so that it can be used by all the editors, not just the timeline editor.

The text event now uses the character resource. Thus the import and export had to be ajusted a bit. When importing, the text event will GUESS the resource, currently based on it's file name. We could change this to account for the actual name property. The guessing code is in the DialogicUtil, so that other events can use it easily as well. 

There has also been a developement on the ResourcePicker (as the character was a string previously). The resource picker allows to search and shows suggestions (filtered only by the file name currently, we might want to account for display name as well). Hitting enter or clicking somewhere else will act as if the first suggestion was selected. You can also drag and drop resources on the picker.

* Add more events to importer/loader

- Moved all currently usable events into addons/dialogic/Resources/Events (and adjusted all references to use this path)

- Added a Comment event (starts with a # in the text format)
- Added an initial version of a Character event to expand the event builder function:
     -> you can now create selectors in the event
     -  there is also a probably working number selector for the events
     - In the text format the Character event currently starts with either "Join", "Leave" or "Update" followed by a character name. Other settings aren't yet implemented.

- Created a "default_event" in addons/dialogic/Resources. This is like a blueprint just containing all necessary methods and some explenation of what they do. This should make it easier to create new events, but we'll see.

- All events now have another static method called `is_valid_event_string` that should return true if the given string should be interpreted as event the method was called on. 
     - the text event always returns true but is asked last
     - the comment event for example checks if the string starts with a # 

Maybe we'll want to change the "one line per event" approach of saving, idk, but until we do, I've decided to convert all \n breaks into <br> characters on save and vice versa on load, so we can have multiple lines in a text event. This isn't nice to write in the text format tho, so who knows how it will end up.

Also renamed all event scripts to start with 'event_', cause then they are in one place in the scripts list and it's easier to see, what they are (in contrast to just text.gd).