Event and timeline improvements for 2.0  (#908)
* Add portrait to character and text event

Also had to update the ResourcePicker to work with portraits. 

I've also added a new method that "Settings Scenes" (those that show a property of the event) can implement called `react_to_change`. If they have a event_resource variable, it will also be set to the one of the event so you can check for other properties. This allows the "portrait picker" to hide if there is no character selected.

Also added that Documentation links are actually opened in the browser. Only the text event has one yet.

Also moved the `identify_event()` function into the DialogicUtil, so it can be used by other scripts as well.

* Make sure character portraits are saved

Currently they are saved as a dictionary. This might change depending on our implmentation of portraits.

The two examples characters have been updated as well as the timelines.

Also fixed the tooltip of the Add Character button.

* Cleanup and Fixes for the TimelineEditor

Removes a number of unnecessary variables and functions of the TimelineEditor.

Changes the `get_events_indexed()` and `add_events_indexed()` and `copy_selected_events()` so copying and duplicating work again.