Add DialogicGameHandler autoload, DialogicDisplay nodes, Event execution
# DialogicGameHandler aka Dialogic singleton
Adds the DialogicGameHandler.gd as an Autoload called Dialogic.

It mainly has the start_timeline() and handle_next_event() functions, as well as methods to update the DialogicDisplay nodes:
- update_dialog_text(text) -> will call all nodes in the `dialogic_dialog_text` group
- update_name_label(text, color) -> will call all nodes in the `dialogic_name_label`group
- update_portrait(character, portrait, position, z_index, animation, etc.) isn't yet implemented...

The autoload is automatically added when the plugin is enabled, so you might have to refresh it.
 
# Dialogic DisplayNodes
There are currently 2 such nodes, the DialogicDialogTextDisplay and the DialogicNameLabelDisplay.
They have a couple of functions each to make them work and are automatically added to the mentioned groups above. 

## DefaultDialogNode
I've added a DefaultDialogNode scene that has instances of both of the above and (right now) autostarts the Chapter1 timeline.

# Event execution
I also had to add the actuall exectuion to the events, but only really did so for the Text event so far. I had to change the "continue_at_end" properties of the events and add finish() to comment and character, so they are basically skipped right now. 
Events get a reference to the dialogic singleton before executing, and then just call the "update" functions of that script (for an example see event_text._execute()).