Questions and Choices 2.0 (#919)
* Jowans Update

# Timeline Preview and Settings File 
There is now a working preview button in the toolbar. To make this work I had to create a settings file again. It's saved as "res://dialogic_settings.cfg" and managed by the DialogicUtil. This might change in the future, but I needed to save the currently edited resource. The preview does only work with saved timelines.

# Choice event 
There is now a choice event. It has nothing but a text input right now. In the text version, choices begin with "-". Choices automatically add a DialogicEndBranchEvent, that is displayed different then the other events, cannot be moved above the choice (if the choice is moved below it's end part, the end is moved as well).
You cannot delete the end node, but deleting the choice will delete the end node as well.
Note: There is currently still some trouble with copy/paste and undo regarding the End node. I will try to fix this soon, if this update is liked.

There is no question event, instead any Text event can start a question.

# ChoiceButton display node
To display choices, there is a new ChoiceButton scene in the Display folder. You can set a button index that will correspond to the choices (first choice get's shown on buttons with button_index == 1, etc.).

There are six choice buttons implemented on the DefaultDialogNode. Multiple functions to manage these buttons have been added to the DialogicGameHandler


# Event class names
All events now have a class name. This means they are easy to recognize in scrips like this:`node.resource is DialogicTextEvent`. 

It also means creating a timeline in script like this is relatively easy:
```	var timeline = DialogicTimeline.new()
	var event1 = DialogicTextEvent.new()
	event1.Text = "This is super cool."
	event1.Character = load("res://timelines/Emilio.dch")
	timeline.add_event(event1)
	DialogicGameHandler.start_timeline(timeline)
```

# Indentation on timelines
When timelines get saved as text, they are getting indented with tabs. I've decided not to indent choices (only what comes after), just because it looked better to me, but we can change that. This is how it looks in text format right now:
```# A cool timeline with two choices
Emilio (Sad): It's quite sad, we cannot use the Tree node.<br>What should we do instead?
- Do everything custom
	John (Nice): Do everything custom! It's the right thing to do!
- Give up
	John (Slick): Just give up. There is nothing we can do

Emilio (Happy): That's a good idea!
```

# Small updates
- Text event will only add brackets in text version if there is a portrait set 
- Test timelines updated

* Fix parsing bugs

The end branch and choice events didn't really work like expected a lot of the times (especially with nested questions). It should have improved significantly now.