Add elif and else mode to conditions (#922)
Adds a ConditionType to the condition event that is either IF, ELIF, or ELSE.

Like before most of the acutal LOGIC is done by the end branch event.

You cannot add elif and else events manually (in the editor), only by using the buttons on the end branch events of if and elif events.

Currently the else event still shows an input for a condition (even though it is ignored). This will change when I add the simplified condition picker.