DialogNode.gd refactoring
After trying to make some custom choice menu I realized the code is not exactly easy to read and I heard the owner of the project on one of his youtube videos saying that the new direction of Dialogic is to make it more accessible for extension.

That said I decided to start cleaning the code and refactoring on my way to understand the workings of the plugin; Here's the first part of that.

Some highlights:
+ Standarized style to more closely follow GDScript's suggested styling (did what I could without changing the names of variables and/or functions that could be being used outside the script as to not break compatibility)
+ + var example : type -> var example: type
+ + enum state -> enum STATE
+ + Spaced static dictionaries
+ + Separated some 'magic' values to constants
+ + Fixed mismatched indentation labels on empty lines
+ Replaced continuous invocations of $ operator to onready variables
+ Removed superfluous comments
+ Standardized comments spelling (Capitalization, spacing, maximum line length)
+ Separated obtuse functionality to their own functions (Increase modularity, reduce function line count, make methods more legible, makes functionality easier to find)
+ Removed unused update_custom_events() function
+ Rephrased some comments to better explain the code
+ Removed instances of internal developer banter from the code (Keep in the commits guys)
+ Replaced long chains of elifs to match statements
+Separated event handling to a function per type (way more legible and easier to modify)
+ Removed commented out code
+ Separated option creation to it's own _create_option(options) function, this will make overriding this functionality (like in my case) possible, further tweaking pending

There are still many very, very, very long functions that need to be decomposed further.
Some of the functionality that DialogNode.gd takes care of should be moved to children nodes, specially the Control node resizing/positioning, that behavior should be responsibility of the children nodes that have a script attached to them.

This same process needs to be applied to the rest of the scripts of the addon, but one step at a time.