Add Character Editor
# Character editor
This re-implements the character editor from 1.x, but with a few changes. 
Scale, offset and mirror are now per-portrait, although I added a new `Main scale` setting that affects all portraits. The "real-size-preview" is a bit more advanced. It has an origin point now and actually displays the portrait offset. It's the new default view. I believe for everything except pixel-art games, this is the view you wan't to use.

There have been a couple of few changes to the timeline editor and toolbar to make them behave when the character editor is opened.