New events and folder movement (#927)
# Folder movement
Moved the images and localization folders into the Editor folder. This caused a lot of autmatic changes, but it seemed to work all fine.

# New events
## Signal event
The new signal event has only a argument setting again. I've added a signal to the dialogicgamehandler called signal_event, because I hope this might be easier understandable than dialogic_signal, but we have to see (renaming this is SUPER easy).
*Text format: `Emit Signal "MyArgument"`*

##  Background event + Background Display
The background event has a input field for a path to a texture or scene. The Background Displaynode is a TextureRect that will either load the texture, or add the scene as it's child. Maybe this isn't optimal, then we can change it oc.
*Text format: `Change Background "res://icon.png"`*

## Change Timeline event
This event has a simple resource picker to select the timeline. The exectution is very simple as well. 
*Text format: `Change Timeline "res://timeline/Chapter1.dtl"` OR*
*Text format: `Change Timeline "Chapter1"` works both, although the latter has to guess the resource* 

## End Timeline (or Close Dialog, idc)
This event has no settings for now.
In execution it will move all characters out, and stop the timeline execution.

# Bug fixes
- The execution of the wait event and text event had been a bit buggy, because the dialog state wasn't correctly set. This should be fixed now.
- the colors of the other events have been changed to what they were in 1.x (except condition, because I want a different color for that)