Auto detect choice button order + empty choices fix
# Auto detect choice button index
The default for the choice button (display node) is now -1 meaning the choices will be distributed to the members of the group in tree order.

# Fixed empty choices issues
If you didn't add events between a choice and it's end event, this resulted in incorrect loading of the timeline (it didn't detect the end branches as there was no change in indentation). This is fixed now and should also be fixed for conditions.

# Other things
- the add elif and add else buttons are now hidden on choices