Add Goto and Label event
This reimplements the goto and label functionality. GoTo only allows to jump inside the timeline, while change timeline now has a label settings as well, that allows to start the given timeline at a specific point.

Alternatively you can also specify a label when starting a timeline in code by doing DialogicGameHandler.start_timeline(resource, label). The label is optional though.

The event icons are a bit to big. Emilio should fix that.