2.0 Jump + Label, Event Button Sorting (#934)
* Add Goto and Label event

This reimplements the goto and label functionality. GoTo only allows to jump inside the timeline, while change timeline now has a label settings as well, that allows to start the given timeline at a specific point.

Alternatively you can also specify a label when starting a timeline in code by doing DialogicGameHandler.start_timeline(resource, label). The label is optional though.

The event icons are a bit to big. Emilio should fix that.

* Implement event button sorting and class names  

Added class names to the events that were missing one. 

Implemented button sorting again.

* Make sure events don't save empty/null properties

* Unify Change Timeline and GoTo into Jump event

The new jump event is basically a Change timeline event with a label selector. If no timeline is selected (or the current one), it will just jump to a label with the name inside the current timeline.

* Make ResourcePicker allow empty values