Implement translation support
# Translation support
## Frontend
### Translation settings
There is a translation tab in the settings editor. It allows to initialize translation files and id's, and then remove them again.

### Implementation on Events
Translations can be implemented on events with the new functions
*To be overridden:*
- can_be_translated() -> bool
- get_original_translation_text() -> String
*To be used:*
- get_translated_text()
*The text event and choice event are implemented, and can be used as reference for further implementation. It's really simple.*

### Translation id
Once translations are initialized, every translatable event gets a unique translation id. This id will be stored in front of the event AUTOMATICALLY (so no way around it) in the format `<xxxx>[event starts here]`.

## Backend
### Translation initialization
Once you hit the initialize button, all timeline files will be opened and all events looked at. A translatable file will get a translation_id and that id, as well as the original text will be added to a csv file.
Also all csv files that get created are added to the projects translation file list.

### Translation updating
If translation is enabled, when you save a timeline in editor, it will change all the texts of the original language, and add new events to the translation file. This means you shouldn't make big changes to your texts after initializing, but can easily edit small errors.

### Event saving/loading changes (for translation_id)
To make it, so the translation_id saving/loading, doesn't have to be implemented by the user (for every event), there is a new layer of methods for save/load on the event resource. This doesn't change anything about how to use the older functions, it just means we can enforce certain things (like in this case saving and loading the <id> in front of the event).

# Other
- Added a helper to the DialogicUtil called get_project_setting, that allows to specify a default value
- edited the editor translation_service.gd so it allows for testing editor scenes
- made sure the MultilineText editor of the text event updates its size on load
- tried adding a title mode to the TLabel, but failed...