Add condition to choice
Adds a condition input to the choice event. UI is very bad but functionality is there. You can also choose, what to do if the condition isn't met (Default/Hide/Disable). Default is currently just the same as Hide, but I want to add a setting for this later.

In the text format you can write the conditions like this:
```
Emilio: Wow, some options are disabled!
- This is disabled [if false] [else disable]
- This might be disabled [if some_cool_condition] [else disabeld]
- This will be hidden [if false] [else hide]
- This is always visible
```

# Other
- moved the condition execution into a function of the dialogic_game_handler, because now multiple events (and the game_handler itself) use it.