2.0 Implement Subsystems for code/logic separation (#975)
* Implement Subsystems + DGH -> Dialogic + Portraits Subsystem

# Introducing Subsystems
There is now a way to implement subsystems to the DialogicGameHandler.
Events have a new function to override:
`get_required_subsystems()`
They have to return them as a list of [name, script_path]'s.

On game start, the DGH will load all the required subsystems. You can then access them by doing `Dialogic.MySubsystem.subsystem_function()`.

Subsystems should (for ease of use) inherit the DialogicSubsystem class.

Subsystems have access to the DGH in a variable called `dialogic`. They can also override 2 methods:
- `clear_game_state()` will be called on game start and before a new save is loaded
- `load_game_state()` will be called when a new save is loaded

# Implementing a Portrait subsystem
Moved all the portrait related methods from the DGH into the new Portrait subsystem.

(Made necessary changes in the character and text event).

# Renamed DialogicGameHandler singleton to Dialogic
The script itself is not renamed and in comments it's often still referred to as DialogicGameHandler which (imo) helps to know what script is meant. 
If you disagree, we can of course also rename the script. But imo saying: `# This is called by Dialogic to do x.` is much less helpfull then `# This is called by the DialogicGameHandler to do x.`

I've also renamed multiple helper variables that reference the DGH to dialogic, because it's just so much shorter.

# Other
- decided to remove the current_state_info helper methods. They made things just more complicated imo.

* Implement more Subsystems

# Subsystems
Adds Subsystems for
- Background
- Audio
- Themes
- Text
- Choices

# Other
- added a clear function to the DialogicGameHandler
- added get_full_state() and load_full_state() methods to the DialogicGameHandler
- fixed a bug with choices, because the EndBranch event checked for "Condition" in event (to determine if it was a Condition event) but Choice event also has "Condition" now.