Move variable stuff into subsystem + Subsystem settings
# Variables subsystem
All the variable-related methods are now in the Variables_Subsystem! This makes the DGH cleaner once again.

# Subsystems can have settings editors
In the `get_required_subsystems()` method of events you can now add a third element to each subsystem specifying the settings editor scene (a path). This scene is added as a tab to the settings editor.

The variables editor is added as such a subsystem editor for example.

pros:
- makes the settings editor cleaner
- makes adding custom settings for "dialogic plugins" easier