2.0 Variables improvement (+ Settingeditor improvements) (#984)
* Add dialogic variables

# Dialogic variables
You can now create and manage dialogic variables from the settings. They can be grouped in a folder structure.

You can now reference them in texts like {MyVariable} or {MyFolder.MySubfolder.MyVariable}.

You can set them with the Set Variable event.

You can get and set them by writing Dialogic.MyVariable = "Hello" or Dialogic.MyFolder.MySubgolder.MyVariable += 10

# Variable editor
The variable editor in the settings supports drag and drop for folders and  variables.

* Move variable stuff into subsystem + Subsystem settings

# Variables subsystem
All the variable-related methods are now in the Variables_Subsystem! This makes the DGH cleaner once again.

# Subsystems can have settings editors
In the `get_required_subsystems()` method of events you can now add a third element to each subsystem specifying the settings editor scene (a path). This scene is added as a tab to the settings editor.

The variables editor is added as such a subsystem editor for example.

pros:
- makes the settings editor cleaner
- makes adding custom settings for "dialogic plugins" easier

* Add _set() ignore for subsystems

Had to add this, otherwise it would throw an error everytime you tried setting the value of an subsyste (like Dialogic.VAR.MyVariable = 20)

Also corrects a mistake in the info text of the variable editor.

* Small Variables Editor improvements

- makes the serach bar more visible
- makes the serach bar look for default values
- renames folders to groups
- makes duplicate names show in a warning color