Exposed Letter Written signal
Added new signal 'letter_displayed', which is emitted whenever a letter is written on the text bubble. This hooks into the existing 'letter_written' signal that plays the typing audio, and bubbles up further with the char that is written (including spaces).

This can be used to build a buffer and react to text as it displays on the screen, or time lip flaps, etc.