Make Display nodes into classes
Removes the display folder and the display scenes. Instead the new Display classes are in their appropriate folders. Also they are classes that extend nodes, so they can be added with the Add Node dialog. (just search dialogic)

Had to remake big parts of the DefaultDialogNode.

From my tests all things should still work, but it's not impossible that I missed something.