2.0 Settings (#1012)
* Implement Choice Settings

- autofocus first choice
- choice delay
- auto hotkeys (1-9)
- default false behaviour

* Text settings + Input Handler

# Text settings
- default speed
- input action (defaults to dialogic_default_action that is now added once the plugin is enabled)
- skippable
- autocolor names
- autocontinue
- autocontinue_delay

# Input Handler
The DialogicInputHandler is a very simple node that has to be added to a dialog scene like the display nodes (unless you want to manage calling the next event yourself). 

It just uses the input action defined in the settings and also cares for the skippable setting.

# Other small changes/bugfixes
- You couldn't use bbcode anymore, so I had to change some regexes a bit.
- writing just [speed] without a parameter will now reset to the current default speed.

* Portrait settings

# RPG portrait mode
Previously named single portrait mode.

# Renamed animation settings
They are now in dialogic/animations/.

* General settings

Added
- custom events folder setting

Moved
- color pallet to dialogic/editor/
- current_timeline to dialogic/editor
- character_preview_mode to dialogic/editor

* Small changes

Added a title mode to the TLabel. It will have a slighlty grey stylebox behind it in this mode.

Moved all choices settings into dialogic/choices/.

* re-adding the plugin icon images

* Fixing the preview timeline setting path

Co-authored-by: Emilio Coppola <coppolaemilio@protonmail.com>