Improve Resource Picker
# Icons
You can now specify an icon for the resource picker.
You can do that also in the info dict in add_header_edit() or add_body_edit().
The keys are 'icon' for a texture or 'editor_icon' for an array like ['Animation', 'EditorIcons'] (because get_icon() is impossible in resources)

- icons have been added to the character, portrait, animation and timeline pickers.

# Dropdown / Empty suggestions
If the search is empty or you click the new drop down button, the first 12 suggestions will show up. This is an arbitrary number that could be changed. If there are more then 12 suggestions a "..." will display at the bottom of the panel.
- many suggestion functions had to be updated to allow an empty search.