2.0 Set Variable Improvements (#1032)
* Set Variable Improvements

# Set Variable
- now has a operator selection (allows "to be", "to itself plus", "to itself minus", "to itself multiplied by" and "to itself divided by")
- now has suggestions for built-in dialogic variables (both on the variable and value picker)

# Complex Picker
I renamed the ResourcePicker to ComplexPicker as it is not used for resources only, but for pretty much anything you want. I've also made small improvements to other input fields:

- hide empty labels on OptionSelector
- make OptionSelector Popup Panel match other menu panels
- make default size for SingleLineText smaller
- hide empty labels on SingleLineText

* Fix ComplexPicker visually trimming variable path because of PrettyName

The Complex Picker was using PrettyName all the time but variables such as Player.Health were not shown correctly because it was thinking .Health was the extension and didn't show it.
Now you can disable pretty_name on the complex picker.

* Fix ComplexPicker resetting values

Hope this fixes it. It seemed to do so for me.

* Make suggestions toggle

* Add Random Int option and Special Writing Method

# Random Int
You can now select a random integer again to be used as the value in the Set Variable event. You can set a min and max value as well.

# Special Text format for Set Variable Event
The Set Variable event is now saved like this:
```
VAR PlayerName = Jonny
VAR MyAutoload.health += 20
VAR Player.Stats.Strength -= [random="True" min="10" max="20] # if you set a value AND random, the value will be ignored
```

# Others
- added a condition option to the add_header_label() method